requirements.txt
hydra-core
dvc
pytest
great_expectations
pandas
scikit-learn
seaborn
matplotlib
jupyter
numpy
zenml[server]
apache-airflow==2.7.3
feast==0.35.0
mlflow==2.14.1
evidently==0.4.23
Flask==2.2.5
flask-session2==1.3.1
Flask-SQLAlchemy==2.5.1
psycopg2-binary==2.9.9
SQLAlchemy==1.4.41
shyaml
mlflow
giskard==2.14.0
gradio==3.50.2
kaggle